struct: recreated a competitive_programming directory to host code related to advanced coding challenges related to ds+algo. -@iamserda
