--- longest_common_prefix/main.py
@@ -1,16 +1,0 @@
-from local_tests import main
-
-
-
-"""WRITE YOUR SOLUTION BELOW"""
-
-
-def longest_common_prefix(string_list: list[str]) -> str:
-    """
-        :type strs: List[str]
-        :rtype: str
-        """
-
-
-if __name__ == "__main__":
-    main.unittest.main()
